+++ apps/etterlatte-saksbehandling-ui/client/src/components/oppgavebenk/MinOppgaveliste.tsx
@@ -1,9 +1,9 @@
 import React, { Dispatch, SetStateAction } from 'react'
 import { VelgOppgavestatuser } from '~components/oppgavebenk/oppgaveFiltrering/VelgOppgavestatuser'
-import { Oppgavelista } from '~components/oppgavebenk/Oppgavelista'
+import { Oppgaver } from '~components/oppgavebenk/oppgaver/Oppgaver'
 import { oppdaterFrist } from '~components/oppgavebenk/utils/oppgaveutils'
 import styled from 'styled-components'
-import { OppgaveFeilWrapper } from '~components/oppgavebenk/OppgaveFeilWrapper'
+import { OppgaveFeilWrapper } from '~components/oppgavebenk/components/OppgaveFeilWrapper'
 import { OppgaveDTO } from '~shared/api/oppgaver'
 import { Result } from '~shared/api/apiUtils'
 import { Filter } from '~components/oppgavebenk/oppgaveFiltrering/oppgavelistafiltre'
@@ -41,7 +41,7 @@ export const MinOppgaveliste = (props: {
             }}
           />
         </ValgWrapper>
-        <Oppgavelista
+        <Oppgaver
           oppgaver={minsideOppgaver}
           oppdaterFrist={(id: string, nyfrist: string, versjon: number | null) =>
             oppdaterFrist(setMinsideOppgaver, minsideOppgaver, id, nyfrist, versjon)
