+++ apps/etterlatte-saksbehandling-ui/client/src/components/oppgavebenk/Oppgavelista.tsx
@@ -1,80 +1,68 @@
-import { Alert } from '@navikt/ds-react'
+import { Oppgaver } from '~components/oppgavebenk/oppgaver/Oppgaver'
+import { FilterRad } from '~components/oppgavebenk/oppgaveFiltrering/FilterRad'
+import { Filter, filtrerOppgaver } from '~components/oppgavebenk/oppgaveFiltrering/oppgavelistafiltre'
+import { OppgaveFeilWrapper } from '~components/oppgavebenk/components/OppgaveFeilWrapper'
 import { OppgaveDTO } from '~shared/api/oppgaver'
-import React, { ReactNode, useEffect, useState } from 'react'
-import { OppgaverTable } from '~components/oppgavebenk/oppgaverTable/OppgaverTable'
-import { PagineringsKontroller } from '~components/oppgavebenk/PagineringsKontroller'
-import {
-  hentSorteringFraLocalStorage,
-  OppgaveSortering,
-  sorterFnr,
-  sorterFrist,
-} from '~components/oppgavebenk/oppgaverTable/oppgavesortering'
+import { Result } from '~shared/api/apiUtils'
 import { Saksbehandler } from '~shared/types/saksbehandler'
-import { hentPagineringSizeFraLocalStorage } from '~components/oppgavebenk/utils/oppgaveutils'
+import { Dispatch, SetStateAction } from 'react'
 
-export interface oppgaveListaProps {
-  oppdaterTildeling: (oppgave: OppgaveDTO, saksbehandler: string | null, versjon: number | null) => void
-  oppgaver: OppgaveDTO[]
-  oppdaterFrist: (id: string, nyfrist: string, versjon: number | null) => void
+export const Oppgavelista = (props: {
+  hovedsideOppgaver: OppgaveDTO[]
+  hentHovedsideOppgaverAlle: () => void
+  hovedsideOppgaverResult: Result<OppgaveDTO[]>
+  gosysOppgaverResult: Result<OppgaveDTO[]>
+  hentHovedsideOppgaver: (oppgavestatusFilter: Array<string>) => void
+  hovedsideFilter: Filter
+  setHovedsideFilter: Dispatch<SetStateAction<Filter>>
   saksbehandlereIEnhet: Array<Saksbehandler>
-  totaltAntallOppgaver?: number
-  erMinOppgaveliste: boolean
-}
+  oppdaterSaksbehandlerTildeling: (oppgave: OppgaveDTO, saksbehandler: string | null, versjon: number | null) => void
+}) => {
+  const {
+    hovedsideOppgaver,
+    hentHovedsideOppgaverAlle,
+    hovedsideOppgaverResult,
+    gosysOppgaverResult,
+    hentHovedsideOppgaver,
+    hovedsideFilter,
+    setHovedsideFilter,
+    saksbehandlereIEnhet,
+    oppdaterSaksbehandlerTildeling,
+  } = props
 
-export const Oppgavelista = ({
-  oppdaterTildeling,
-  oppgaver,
-  oppdaterFrist,
-  saksbehandlereIEnhet,
-  totaltAntallOppgaver,
-  erMinOppgaveliste,
-}: oppgaveListaProps): ReactNode => {
-  const [sortering, setSortering] = useState<OppgaveSortering>(hentSorteringFraLocalStorage())
+  const mutableOppgaver = hovedsideOppgaver.concat()
 
-  const sortertFrist = sorterFrist(sortering.fristSortering, oppgaver)
-  const sorterteOppgaver = sorterFnr(sortering.fnrSortering, sortertFrist)
-
-  const [page, setPage] = useState<number>(1)
-  const [rowsPerPage, setRowsPerPage] = useState<number>(hentPagineringSizeFraLocalStorage())
-
-  let paginerteOppgaver = sorterteOppgaver
-  paginerteOppgaver = paginerteOppgaver.slice((page - 1) * rowsPerPage, page * rowsPerPage)
-
-  useEffect(() => {
-    if (paginerteOppgaver.length === 0 && oppgaver.length > 0) setPage(1)
-  }, [paginerteOppgaver, oppgaver])
+  const filtrerteOppgaver = filtrerOppgaver(
+    hovedsideFilter.enhetsFilter,
+    hovedsideFilter.fristFilter,
+    hovedsideFilter.saksbehandlerFilter,
+    hovedsideFilter.ytelseFilter,
+    hovedsideFilter.oppgavestatusFilter,
+    hovedsideFilter.oppgavetypeFilter,
+    hovedsideFilter.oppgavekildeFilter,
+    mutableOppgaver,
+    hovedsideFilter.fnrFilter
+  )
 
   return (
-    <>
-      {paginerteOppgaver && paginerteOppgaver.length > 0 ? (
-        <>
-          <PagineringsKontroller page={page} setPage={setPage} antallSider={Math.ceil(oppgaver.length / rowsPerPage)} />
-
-          <OppgaverTable
-            oppgaver={paginerteOppgaver}
-            oppdaterTildeling={oppdaterTildeling}
-            erMinOppgaveliste={erMinOppgaveliste}
-            oppdaterFrist={oppdaterFrist}
-            saksbehandlereIEnhet={saksbehandlereIEnhet}
-            setSortering={setSortering}
-          />
-
-          <PagineringsKontroller
-            page={page}
-            setPage={setPage}
-            antallSider={Math.ceil(oppgaver.length / rowsPerPage)}
-            raderPerSide={rowsPerPage}
-            setRaderPerSide={setRowsPerPage}
-            totalAvOppgaverTeksts={`Viser ${(page - 1) * rowsPerPage + 1} - ${
-              (page - 1) * rowsPerPage + paginerteOppgaver.length
-            } av ${oppgaver.length} oppgaver ${
-              totaltAntallOppgaver ? `(totalt ${totaltAntallOppgaver} oppgaver)` : ''
-            }`}
-          />
-        </>
-      ) : (
-        <Alert variant="info">Ingen oppgaver</Alert>
-      )}
-    </>
+    <OppgaveFeilWrapper oppgaver={hovedsideOppgaverResult} gosysOppgaver={gosysOppgaverResult}>
+      <>
+        <FilterRad
+          hentAlleOppgaver={hentHovedsideOppgaverAlle}
+          hentOppgaverStatus={(oppgavestatusFilter: Array<string>) => hentHovedsideOppgaver(oppgavestatusFilter)}
+          filter={hovedsideFilter}
+          setFilter={setHovedsideFilter}
+          alleOppgaver={hovedsideOppgaver}
+        />
+        <Oppgaver
+          oppgaver={filtrerteOppgaver}
+          oppdaterTildeling={oppdaterSaksbehandlerTildeling}
+          oppdaterFrist={() => {}}
+          totaltAntallOppgaver={hovedsideOppgaver.length}
+          erMinOppgaveliste={false}
+          saksbehandlereIEnhet={saksbehandlereIEnhet}
+        />
+      </>
+    </OppgaveFeilWrapper>
   )
 }
