+++ apps/etterlatte-saksbehandling-ui/client/src/components/oppgavebenk/oppgaveModal/GosysOppgaveModal.tsx
@@ -2,7 +2,7 @@ import { BodyShort, Button, Heading, Label, Modal } from '@navikt/ds-react'
 import styled from 'styled-components'
 import { EyeIcon } from '@navikt/aksel-icons'
 import { useContext, useState } from 'react'
-import { OppgavetypeTag, SaktypeTag } from '~components/oppgavebenk/Tags'
+import { OppgavetypeTag, SaktypeTag } from '~components/oppgavebenk/components/Tags'
 import { formaterFnr, formaterStringDato } from '~utils/formattering'
 import { OppgaveDTO } from '~shared/api/oppgaver'
 import { ConfigContext } from '~clientConfig'
