+++ apps/etterlatte-saksbehandling-ui/client/src/components/oppgavebenk/oppgaverTable/OppgaverTableRow.tsx
@@ -3,9 +3,9 @@ import { Table } from '@navikt/ds-react'
 import { erOppgaveRedigerbar, OppgaveDTO } from '~shared/api/oppgaver'
 import { formaterStringDato } from '~utils/formattering'
 import { FristWrapper } from '~components/oppgavebenk/frist/FristWrapper'
-import SaksoversiktLenke from '~components/oppgavebenk/SaksoversiktLenke'
-import { OppgavetypeTag, SaktypeTag } from '~components/oppgavebenk/Tags'
-import { HandlingerForOppgave } from '~components/oppgavebenk/HandlingerForOppgave'
+import SaksoversiktLenke from '~components/oppgavebenk/components/SaksoversiktLenke'
+import { OppgavetypeTag, SaktypeTag } from '~components/oppgavebenk/components/Tags'
+import { HandlingerForOppgave } from '~components/oppgavebenk/components/HandlingerForOppgave'
 import { FristHandlinger } from '~components/oppgavebenk/frist/FristHandlinger'
 import { VelgSaksbehandler } from '~components/oppgavebenk/tildeling/VelgSaksbehandler'
 import { OPPGAVESTATUSFILTER } from '~components/oppgavebenk/oppgaveFiltrering/oppgavelistafiltre'
